refactor: split Explore into custom hooks and replace timeout with onLoading
Split Explore.tsx (660 lines -> 335 lines) into three custom hooks:
- useExploreVisibility: visibility/scope logic
- useExploreControllers: controller subscriptions
- useExploreData: data loading and onLoading subscription

Replace fragile 5s setTimeout fallback with proper onLoading callbacks
from nostrverseWritingsController and nostrverseHighlightsController.
